requirements.txt
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.14.6
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
duckduckgo_search==6.1.12
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.8.0.post1
fastapi==0.104.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.43.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.3
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.0
grpcio-status==1.65.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel==6.25.2
ipython==8.16.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.12.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.15
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.83
layoutparser==0.3.4
lxml==4.9.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
multitasking==0.0.11
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2
nltk==3.8.1
numpy==1.26.1
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.10
opencv-python==4.10.0.84
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.6
packaging==23.2
pandas==2.1.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
peewee==3.17.6
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pikepdf==9.0.0
Pillow==10.1.0
pillow_heif==0.17.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
portalocker==2.10.0
prometheus-client==0.17.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.6
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==13.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.9.1
Pygments==2.16.1
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyrate-limiter==3.6.1
pyreadline3==3.4.1
pyreqwest_impersonate==0.4.9
pytesseract==0.3.10
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.0
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2023.3.post1
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.3
sec-edgar-downloader==5.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.49
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
streamlit==1.32.2
sympy==1.12
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.28.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
unstructured==0.14.10
unstructured-client==0.24.0
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.23.2
watchdog==4.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
websocket-client==1.6.4
wikipedia==1.4.0
wordcloud==1.9.2
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
yfinance==0.2.40
youtube-search==2.1.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
